requirements.txt
requests >= 2.0
future
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.9
Flask==1.0
future==0.16.0
gunicorn==19.8.1
httpie==0.9.9
idna==2.6
itsdangerous==0.24
Jinja2==2.10
line-bot-sdk==1.5.0
MarkupSafe==1.0
pipenv==11.10.0
Pygments==2.2.0
PyMySQL==0.8.0
requests==2.18.4
urllib3==1.22
virtualenv==15.2.0
virtualenv-clone==0.3.0
Werkzeug==0.14.1
psycopg2>=2.7,<2.8 --no-binary psycopg2
